Update local gpu solver
