+++ evals/solvers/providers/qwen/qwen2audio_local_gpu_solver.py
@@ -16,6 +16,7 @@ from concurrent import futures
 from concurrent.futures import ProcessPoolExecutor
 from typing import Any, Callable, Dict, List, Optional, TypeVar
 from evals.solvers.solver import Solver, SolverResult
+from evals.solvers.utils import BatchedProcessPoolExecutor
 from evals.task_state import TaskState
 
 SAMPLE_RATE = 16000
@@ -203,127 +204,3 @@ def solver_worker(inputs: List[Dict[str, Any]]) -> List[str]:
     return responses
 
 
-
-T_In = TypeVar("T_In")
-T_Out = TypeVar("T_Out")
-
-
-@dataclasses.dataclass
-class BatchableWorkItem:
-    request: T_In
-    future: futures.Future
-
-
-class BatchedProcessPoolExecutor:
-    def __init__(
-        self,
-        *args,
-        batch_worker_fn: Callable[[List[T_In]], List[T_Out]],
-        max_batch_size: int,
-        max_workers: int = 1,
-        **kwargs
-    ):
-        self.max_batch_size = max_batch_size
-        self.batch_worker_fn = batch_worker_fn
-        self._batch_queue = queue.Queue()
-        self.available_workers = threading.Semaphore(value=max_workers + 1)
-        self.process_pool_executor = ProcessPoolExecutor(
-            *args, max_workers=max_workers, **kwargs
-        )
-        self._batch_thread = threading.Thread(target=self.batch_requests)
-        self._batch_thread.start()
-
-    def submit(self, request: T_In) -> futures.Future:
-        item = BatchableWorkItem(request, futures.Future())
-        self._batch_queue.put(item)
-        return item.future
-
-    def shutdown(self):
-        # shut down the process pool executor
-        self.process_pool_executor.shutdown()
-
-        while not self._batch_queue.empty():
-            try:
-                item = self._batch_queue.get(block=False)
-                if item is not None:
-                    item.future.set_exception(Exception("The pool has already shut down."))
-            except queue.Empty:
-                pass
-
-        # signal the batch thread to stop
-        self._batch_queue.put(None)
-        print("batched process pool executor has shut down.")
-
-    def batch_requests(self):
-        """
-        Batches requests and dispatches them to the process pool executor
-
-        Steps:
-        1. greedily grab items
-        2. dispatch them to ProcessPoolExecutor
-        3. set the results back on the source future
-        """
-        # Wait a bit for the GPUs to be ready and allow the batch_queue to fill up a bit
-        time.sleep(1)
-
-        while True:
-            # We don't wait to rush ahead too fast and fill up the queue with
-            # batch_size=1 requests, but we also don't want any GPUs to be idle.
-            self.available_workers.acquire()
-
-            # greedily grab items
-            work_items: List[BatchableWorkItem] = [self._batch_queue.get()]
-            while len(work_items) < self.max_batch_size:
-                try:
-                    item = self._batch_queue.get(block=False)
-                    work_items.append(item)
-                except queue.Empty:
-                    break
-
-            # When we're done, a None item is added to the queue to signal the end of requests.
-            # We will ignore any existing work_items since the process pool executor
-            # is already shutting down.
-            if work_items[-1] is None:
-                if len(work_items) > 1:
-                    logging.warn(
-                        "There remained work items in the queue when shutting down. The items will be ignored."
-                    )
-                return
-
-            requests = [item.request for item in work_items]
-            task_futures = [item.future for item in work_items]
-
-            # dispatch to the process pool
-            try:
-                result_future = self.process_pool_executor.submit(self.batch_worker_fn, requests)
-            except Exception as e:
-                self._handle_exception(e, task_futures)
-                return
-
-            # add callback for when the result is ready
-            result_future.add_done_callback(_set_results_cb(task_futures, self._handle_exception))
-            result_future.add_done_callback(lambda _: self.available_workers.release())
-
-    def _handle_exception(self, e: Exception, task_futures: List[futures.Future]):
-        """
-        Handles exceptions by simply panicking:
-         * prints the traceback to allow debugging
-         * sets exception on all the task futures
-         * shuts down the executor
-        """
-        print(traceback.format_exc())
-        for f in task_futures:
-            if not f.done():
-                f.set_exception(e)
-        self.shutdown()
-
-
-def _set_results_cb(task_futures: List[futures.Future], handle_exception_cb: Callable):
-    def cb(batch_future: futures.Future):
-        try:
-            for f, r in zip(task_futures, batch_future.result()):
-                f.set_result(r)
-        except Exception as e:
-            handle_exception_cb(e, task_futures)
-
-    return cb
